Rename: type파일 types로 변경
