+++ components/Schedule/types.ts
